server/requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15